pyproject.toml
[tool.poetry]
name        = "mtc-bot"
version     = "0.1.0"
description = "Multi-Tenant Crypto Trading Bot monorepo"
readme      = "README.md"

packages = [
  { include = "bot_core",         from = "src" },
  { include = "data_service",     from = "src" },
  { include = "strategy_service", from = "src" },
  { include = "strategy_dsl",     from = "src" },
  { include = "broker_adapters",  from = "src" },
  { include = "common",           from = "src" },
  { include = "user_service",     from = "src" },
  { include = "api_gateway",      from = "src" },
  { include = "web_ui",           from = "src" },
]

authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python               = ">=3.12,<4"
ta                   = ">=0.11.0,<0.12.0"
optuna               = ">=4.4.0,<5.0.0"
pybit                = ">=5.11.0,<6.0.0"
aiokafka             = "^0.12.0"
# Machine Learning dependencies will be installed separately
# torch                = "^2.0.0"
# torchaudio           = "^2.0.0"
# torchvision          = "^0.15.0"
numpy                = "^1.26.0"
pandas               = "^2.1.0"
scikit-learn         = "^1.3.0"
transformers         = "^4.36.0"
# Time series analysis
einops               = "^0.7.0"
# Data processing
pydantic             = "^2.5.0"
pydantic-settings    = "^2.1.0"
# Visualization
matplotlib           = "^3.8.0"
seaborn              = "^0.13.0"
plotly               = "^5.17.0"
# Other utilities
tqdm                 = "^4.66.0"
PyYAML               = "^6.0.1"
tensorboard          = "^2.15.0"
# GPU monitoring
nvidia-ml-py         = "^12.535.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.3"
mypy = "^1.16.1"
